structure refactor hooks
