+++ src/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './use-element-click-positions';
+export * from './use-element-click-positions/use-element-click-positions'
